[#544] Moved Sync tasks and loaders + fixed previously introduced compilation errors
