+++ app/src/main/java/org/akvo/flow/data/loader/base/AsyncLoader.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2013 Stichting Akvo (Akvo Foundation)
+ *  Copyright (C) 2013-2016 Stichting Akvo (Akvo Foundation)
  *
  *  This file is part of Akvo FLOW.
  *
@@ -14,7 +14,7 @@
  *  The full license text can also be seen at <http://www.gnu.org/licenses/agpl.html>.
  */
 
-package org.akvo.flow.async.loader.base;
+package org.akvo.flow.data.loader.base;
 
 import android.content.Context;
 import android.support.v4.content.AsyncTaskLoader;
