+++ app/src/main/java/org/akvo/flow/ui/fragment/DrawerFragment.java
@@ -44,8 +44,8 @@ import android.widget.Toast;
 import org.akvo.flow.R;
 import org.akvo.flow.activity.SettingsActivity;
 import org.akvo.flow.app.FlowApp;
-import org.akvo.flow.async.loader.SurveyGroupLoader;
-import org.akvo.flow.async.loader.UserLoader;
+import org.akvo.flow.data.loader.SurveyGroupLoader;
+import org.akvo.flow.data.loader.UserLoader;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.data.database.UserColumns;
 import org.akvo.flow.domain.SurveyGroup;
