+++ app/src/main/java/org/akvo/flow/ui/fragment/MapFragment.java
@@ -49,7 +49,7 @@ import com.google.maps.android.clustering.view.DefaultClusterRenderer;
 import org.akvo.flow.R;
 import org.akvo.flow.activity.RecordActivity;
 import org.akvo.flow.activity.SurveyActivity;
-import org.akvo.flow.async.loader.SurveyedLocaleLoader;
+import org.akvo.flow.data.loader.SurveyedLocaleLoader;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.domain.SurveyGroup;
 import org.akvo.flow.domain.SurveyedLocale;
