+++ app/src/main/java/org/akvo/flow/ui/fragment/ResponseListFragment.java
@@ -39,7 +39,7 @@ import android.widget.ListView;
 import org.akvo.flow.R;
 import org.akvo.flow.activity.FormActivity;
 import org.akvo.flow.activity.TransmissionHistoryActivity;
-import org.akvo.flow.async.loader.SurveyInstanceLoader;
+import org.akvo.flow.data.loader.SurveyInstanceLoader;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.domain.SurveyGroup;
 import org.akvo.flow.domain.SurveyedLocale;
