+++ app/src/main/java/org/akvo/flow/ui/fragment/SurveyedLocaleListFragment.java
@@ -47,7 +47,7 @@ import android.widget.Toast;
 
 import org.akvo.flow.R;
 import org.akvo.flow.activity.SurveyActivity;
-import org.akvo.flow.async.loader.SurveyedLocaleLoader;
+import org.akvo.flow.data.loader.SurveyedLocaleLoader;
 import org.akvo.flow.data.database.SurveyDbAdapter;
 import org.akvo.flow.data.database.RecordColumns;
 import org.akvo.flow.data.database.SurveyInstanceColumns;
